backend/pyproject.toml
[project]
name = "mental-health-backend"
version = "0.1.0"
description = "FastAPI + LangGraph mental health chat API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.6.0",
    "motor>=3.6.0",
    "python-dotenv>=1.0.1",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-groq>=0.2.0",
    "httpx>=0.27.0",
    "bcrypt>=4.2.0",
    "PyJWT>=2.9.0",
    "email-validator>=2.2.0",
    "redis[asyncio]>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
